+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [0.1.0] - 2025-01-01
 ### Added
+- SecondsDisabler component
+- FramesDisabler component
+- BecameInvisibleDisabler component
+- AnimationDisabler component
 - PoolSystem component
 - Poolable component
 - PoolableData